[xconfmap] Remove `Validator` and `Validate`
